More beer work

+++ src/java/fantasykingdoms/common/items/ItemEmptyTankard.java
@@ -5,6 +5,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 
 import fantasykingdoms.common.FantasyKingdoms;
+import fantasykingdoms.common.tiles.TileBarrel;
 
 public class ItemEmptyTankard extends BaseItem
 {
@@ -17,10 +18,12 @@ public class ItemEmptyTankard extends BaseItem
 	@Override
 	public boolean onItemUse(ItemStack stack, EntityPlayer player, World world, int x, int y, int z, int side, float px, float py, float pz)
 	{
-		if (world.getBlock(x, y, z) == FantasyKingdoms.blockBarrel)
+		TileBarrel tile = (TileBarrel) world.getTileEntity(x, y, z);
+		if (world.getBlock(x, y, z) == FantasyKingdoms.blockBarrel && tile.getBeerType() != "empty" && tile.getBeerLevel() > 0)
 		{
+			tile.setBeerLevel(tile.getBeerLevel() - 1);
 			player.destroyCurrentEquippedItem();
-			player.inventory.addItemStackToInventory(new ItemStack(FantasyKingdoms.itemEmptyTankard));
+			player.inventory.addItemStackToInventory(new ItemStack(FantasyKingdoms.itemFullTankard));
 			return true;
 		}
